feat: implement endpoint for organization deletion
